feat: 1.http요청 메서드 utility 메서드로 변경 2. 테스트 코드 추가
